lavoro funzionalita delete item e delete all

+++ utente_registrato/cart.php
@@ -1,19 +1,19 @@
-<?php 
-    include("_header.php");
-    include("../include/_db_dal.inc.php");
-    $conn = db_connect();
-    if(!empty($_SESSION["id"])) {
-        $id = $_SESSION["id"];
-        $cart_list = get_products_by_user($conn, $id);
-        $result = mysqli_query($conn, "SELECT * FROM utente WHERE id_u = $id");
-        $user = $row = mysqli_fetch_assoc($result);
-        //echo print_r($cart_list);
-    } else {
-        $_SESSION["login"] = false;
-        $_SESSION["id"] = 0;
-        header("Location: ../auth/login.php");
-    }
-    /*
+<?php
+include("_header.php");
+include("../include/_db_dal.inc.php");
+$conn = db_connect();
+if (!empty($_SESSION["id"])) {
+    $id = $_SESSION["id"];
+    $cart_list = get_products_by_user($conn, $id);
+    $result = mysqli_query($conn, "SELECT * FROM utente WHERE id_u = $id");
+    $user = $row = mysqli_fetch_assoc($result);
+    //echo print_r($cart_list);
+} else {
+    $_SESSION["login"] = false;
+    $_SESSION["id"] = 0;
+    header("Location: ../auth/login.php");
+}
+/*
     $action = isset($_GET["action"]) ? $_GET["action"] : "none";
     $id = isset($_GET["code"]) ? $_GET["code"] : "none";
     $cart_list = [];
@@ -33,37 +33,45 @@
 <div class="cart-container">
     <div class="cart-left">
         <div>
-            <h1><?=$user["nome"]?>'s Shopping Cart</h1>
+            <h1><?= $user["nome"] ?>'s Shopping Cart</h1>
         </div>
-        <?php foreach ($cart_list as $cart_item) { ?>
-        <div class="cart-item">
-            <img src="../assets/img/table_lamp.png" alt="">
-            <div class="cart-item-info">
-                <h3><?=$cart_item["titolo"]?></h3>
-                <div class="quantity">
-                    <input readonly id="<?=$cart_item["idEc"]?>" type="number" min="1" max="90" step="1" value="<?=$cart_item["quantita"]?>">
+        <?php 
+        if(count($cart_list) > 0) {
+            foreach ($cart_list as $cart_item) { 
+        ?>
+            <div class="cart-item">
+                <img src="../assets/img/table_lamp.png" alt="">
+                <div class="cart-item-info">
+                    <h3><?= $cart_item["titolo"] ?></h3>
+                    <div class="quantity">
+                        <input readonly id="<?= $cart_item["idEc"] ?>" type="number" min="1" max="90" step="1" value="<?= $cart_item["quantita"] ?>">
+                    </div>
+                    <div class="cart-item-price" data-price="<?= $cart_item["prezzo"] ?>">
+                        <h1><?= ($cart_item["prezzo"] * $cart_item["quantita"]) ?>$</h1>
+                    </div>
                 </div>
-                <div class="cart-item-price" data-price="<?=$cart_item["prezzo"]?>">
-                    <h1><?=($cart_item["prezzo"]*$cart_item["quantita"])?>$</h1>
-                </div>
+                <button class="cart-item-delete">Delete</button>
             </div>
-            <button class="cart-item-delete">Delete</button>
-        </div>
-        <?php } ?>
+        <?php } }else{ ?>
+            <h1>You've got no items in your cart <a href="../index.php">Shop now!</a></h1>
+        <?php }?>
     </div>
     <div class="cart-right">
+        <?php if(count($cart_list) > 0) { ?>
         <div class="subtotal">
-            <h2>Subtotal (<?=count($cart_list)?> items):</h2>
-            <h1><?=get_total_price($conn, $id)?>$</h1>
+            <h2>Subtotal - <?= count($cart_list) ?> item(s):</h2>
+            <h1><?php echo get_total_price($conn, $id) ."$" ?></h1>
             <button class="cart-buy-button">Buy Now</button>
             <button class="cart-delete-all">Delete Cart</button>
         </div>
+        <?php } ?>
     </div>
 </div>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.4/jquery.min.js"></script>
 <script>
     $(document).ready(function() {
         var subtotal = 0;
+        //create counter and initialize variables
         jQuery('<div class="quantity-nav"><div class="quantity-button quantity-up">+</div><div class="quantity-button quantity-down">-</div></div>').insertAfter('.quantity input');
         jQuery('.quantity').each(function() {
             var spinner = jQuery(this),
@@ -74,12 +82,11 @@
                 max = input.attr('max'),
                 idEc = input.attr('id'),
                 oldValue = parseFloat(input.val());
-                //console.log(spinner.next().text());
-                //console.log($(".subtotal").find("h1").text());
-                subtotal += parseFloat(spinner.next().attr("data-price")) * oldValue;
-                console.log(subtotal);
+            //get total price
+            subtotal += parseFloat(spinner.next().attr("data-price")) * oldValue;
 
             btnUp.click(function() {
+                //update input counter
                 oldValue = parseFloat(input.val());
                 if (oldValue >= max) {
                     var newVal = oldValue;
@@ -88,19 +95,21 @@
                 }
                 spinner.find("input").val(newVal);
                 spinner.find("input").trigger("change");
+                //prepare AJAX request
                 const xhttp = new XMLHttpRequest();
                 xhttp.onload = function() {
                     console.log(this.responseText);
                 }
                 xhttp.open("GET", `update_prod_qty.php?idEc=${idEc}&curval=${newVal-1}&qty=1`);
                 xhttp.send();
+                //update price display
                 spinner.next().html(`<h1>${newVal*spinner.next().attr("data-price")}$</h1>`);
                 subtotal += parseFloat(spinner.next().attr("data-price"));
-                console.log(subtotal);
                 $(".subtotal").find("h1").html(`${subtotal}` + "$");
             });
 
             btnDown.click(function() {
+                //update input counter
                 oldValue = parseFloat(input.val());
                 if (oldValue <= min) {
                     var newVal = oldValue;
@@ -109,44 +118,60 @@
                 }
                 spinner.find("input").val(newVal);
                 spinner.find("input").trigger("change");
+                //prepare AJAX request
                 const xhttp = new XMLHttpRequest();
                 xhttp.onload = function() {
                     console.log(this.responseText);
                 }
-                xhttp.open("GET", `update_prod_qty.php?idEc=${idEc}&curval=${newVal+1}&qty=-1`, true);
+                xhttp.open("GET", `update_prod_qty.php?idEc=${idEc}&curval=${newVal+1}&qty=-1`);
                 xhttp.send();
+                //update price display
                 spinner.next().html(`<h1>${newVal*spinner.next().attr("data-price")}$</h1>`);
                 subtotal -= parseFloat(spinner.next().attr("data-price"));
                 $(".subtotal").find("h1").html(`${subtotal}` + "$");
             });
         });
-
+        //removes all items
         $("button.cart-delete-all").click(function() {
-            $(".cart-left > .cart-item").remove();
+            //initialize variables
+            var cartItem = jQuery(this),
+                input = cartItem.find('input[type="number"]'),
+                idEc = input.attr('id');
+            if(confirm("Do you want to remove all items?")){
+                //prepare AJAX request
+                const xhttp = new XMLHttpRequest();
+                xhttp.onload = function() {
+                    subtotal = this.responseText;
+                    $(".subtotal").find("h1").text(`${subtotal}$`);
+                }
+                xhttp.open("GET", `delete_item.php?idEc=${idEc}&uId=${<?=$id?>}&action=delete_all`);
+                xhttp.send();
+                $(".cart-left > .cart-item").remove();
+            }else{
+                console.log("cancelled");
+            }
         });
-        
-        $(".cart-item").each(function(){
+
+        $(".cart-item").each(function() {
+            //initialize variables
             var cartItem = jQuery(this),
                 input = cartItem.find('input[type="number"]'),
                 btnDelete = cartItem.find(".cart-item-delete"),
-                idEc = input.attr('id'),
-                oldValue = parseFloat(input.val());
-            console.log(idEc);
+                idEc = input.attr('id');
+            //removes one item
             btnDelete.click(function() {
+                //prepare AJAX request
                 const xhttp = new XMLHttpRequest();
                 xhttp.onload = function() {
-                    console.log(this.responseText);
+                    subtotal = this.responseText;
+                    $(".subtotal").find("h1").text(`${subtotal}$`);
                 }
-                xhttp.open("GET", `delete_item.php?idEc=${idEc}`);
+                xhttp.open("GET", `delete_item.php?idEc=${idEc}&uId=${<?=$id?>}&action=delete_one`);
                 xhttp.send();
-                subtotal = jQuery(".quantity").each(function(){
-                    subtotal += parseFloat(cartItem.next().attr("data-price")) * oldValue
-                });
-                $(".subtotal").find("h1").html(`${subtotal}` + "$"); //DA FARE
                 cartItem.remove();
             });
         });
-        
+
     });
 </script>
 <?php include("_footer.php") ?>
